fix player_index that was int and should have been size_t

+++ executables/LUDUMDARE/Ludum43/src/Ludum43Game.h
@@ -64,11 +64,11 @@ public:
 
 	virtual chaos::box2 GetWorldBox() const override;
 
-	float GetPlayerParticleLife(int player_index) const;
+	float GetPlayerParticleLife(size_t player_index) const;
 
-	ParticlePlayer * GetPlayerParticle(int player_index);
+	ParticlePlayer * GetPlayerParticle(size_t player_index);
 
-	ParticlePlayer const * GetPlayerParticle(int player_index) const;
+	ParticlePlayer const * GetPlayerParticle(size_t player_index) const;
 
 protected:
