+++ executables/LUDUMDARE/Ludum44/src/Ludum44Game.h
@@ -73,9 +73,9 @@ public:
 	virtual chaos::box2 GetWorldBox() const override;
 
 
-	ParticlePlayer * GetPlayerParticle(int player_index);
+	ParticlePlayer * GetPlayerParticle(size_t player_index);
 
-	ParticlePlayer const * GetPlayerParticle(int player_index) const;
+	ParticlePlayer const * GetPlayerParticle(size_t player_index) const;
 
 	float GetBuyUpgradeTime() const { return buy_upgrade_time; }
